requirements.txt
django>=4.2.7
psycopg2>=2.9.6
djangorestframework>=3.14.0
django-filter>=23.2
drf-yasg>=1.21.5
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.4
gunicorn
whitenoise
dj-database-url
coverage